Bug fix to callback URI, etc. Also added some logging and the coaster.worker.script and coaster.worker.logging properties
git-svn-id: https://svn.code.sf.net/p/cogkit/svn/trunk/current/src@2756 5b74d2a0-fa0e-0410-85ed-ffba77ec0bde

+++ cog/modules/provider-coaster/src/org/globus/cog/abstraction/coaster/service/job/manager/ScriptManager.java
@@ -27,7 +27,8 @@ public class ScriptManager {
             throw new IOException("Failed to create script dir (" + scriptDir + ")");
         }
         File script = File.createTempFile("cscript", ".pl", scriptDir);
-        script.deleteOnExit();
+        if (! "persistent".equals(System.getProperty("coaster.worker.script")))
+            script.deleteOnExit();
         InputStream is = ScriptManager.class.getClassLoader().getResourceAsStream(SCRIPT);
         if (is == null) {
             throw new IOException("Could not find resource in class path: " + SCRIPT);
